requirements.txt
SQLAlchemy==1.0.12
mysqlclient==1.3.7
pyyaml==3.11